A weekend's worth of merging, in preparation for the Typography switchover.
MOVED
-----

* Win32/Crypt/md5.c -> Common/libs/crypto/md5.c
* Win32/Crypt/md5.h -> Common/libs/crypto/md5.h
* Win32/Crypt/md5main.c -> Common/libs/crypto/md5main.c
* Win32/Crypt/_blowfish.c -> Common/libs/crypto/blowfish.c
* Win32/Crypt/blowfish.h -> Common/libs/crypto/blowfish.h

Third-party cryptography files moved to the new Common/libs/crypto
directory.  In general, third-party code should go under Common/libs, so we
can find it all in one place for updates and license checks.
Common/freetype2 will probably move there soon for the sake of consistency.

MERGED
------

* Win32/Crypt/CryptStream.cpp -> Common/CryptStream.cpp
* Win32/Crypt/CryptStream.h -> Common/CryptStream.h
* Win32/TestSuite/TestCryptStream.cpp -> Common/CryptStreamTests.cpp

Modified to use the portable Path abstraction.  Included our standard key
once in this file, instead of having it in many different headers
throughout the program. Coerced uchar* to char* in several places required
by the fstream API (and some other coercions).

* Win32/FiveL/Parser.cpp -> Common/TParser.cpp
* Win32/FiveL/Parser.h -> Common/TParser.h

Merged in Elizabeth's improved escape-handling code.  Factored out all code
which specifically referred to "card", "header" or "macrodef" forms, and
added a generic API for registering abitrary top-level forms.

* Win32/FiveL/Index.cpp -> Common/TIndex.cpp
* Win32/FiveL/Index.h -> Common/TIndex.h
* NEW: Common/TIndexTests.cpp
* NEW: Common/Scripts/test.scr

Merged TIndex::GetScript from the Macintosh.  Temporarily stopped closing
the TIndexFile in the presence of REDOSCRIPT.  Merged some Macintosh code
for building indices from FSSpecs; this probably doesn't work.  Changed the
Open and Init methods to use the portable Path library (the APIs might be
slightly suboptimal).

* Win32/FiveL/LUtil.cpp -> Common/TDateUtil.cpp
* Win32/FiveL/LUtil.h -> Common/TDateUtil.h

Extracted date-related code from LUtil.*.  Changed wsprintf calls to
sprintf.

* Win32/FiveL/Variable.cpp -> Common/TVariable.cpp
* Win32/FiveL/Variable.h -> Common/TVariable.h

Disabled certain special variables that caused awkward dependencies, and
replaced them with an interface for registering arbitrary special
variables.

MODIFIED
--------

* Common/FileSystem.cpp
* Common/FileSystem.h

Added a RenameFile function, and a GetScriptsDirectory function.  Also
added a ReplaceWithTemporaryFile function, which overwrites an existing
file with a temporary file (someday, we can implement this as an atomic
operation on most operating systems).

* Common/GraphicsTools.h

Added a no-arguments constuctor for Point.

* Common/TString.cpp
* Common/TString.h

Lots of "signed/unsigned comparison" and other warning fixes.

* Common/TStyleSheet.cpp
* Common/TStyleSheet.h

Added full-fledged INCR_X, INCR_Y support!

* Common/Typography.cpp
* Common/Typography.h

Made sure that kerning+advance can never move the drawing cursor backwards.
Fixed warnings.

* Common/fonttools/pngtest.cpp

Added a test of transparent text (just for fun).

KNOWN ISSUES
------------

* Logging code needs to have Mac-specific features merged back in.

* TIndexFile doesn't close the underlying file properly in the presence of
REDOSCRIPT.  What's going on here?

* TParser--and maybe TStream--need to have cross-platform end-of-line
handling.

+++ Common/Typography.h
@@ -280,7 +280,7 @@ namespace Typography {
 		std::wstring mText;
 		std::map<size_t,Style> mStyleRuns;
 		bool mIsBuilt;
-		int mEnd;
+		size_t mEnd;
 
 	public:
 		//////////
@@ -374,7 +374,7 @@ namespace Typography {
 			friend class StyledText;
 
 			const StyledText *mStyledText;
-			int mCurrentPosition;
+			size_t mCurrentPosition;
 			value_type mCurrentValue;
 
 			const_iterator(const StyledText *inStyledText, size_t inPos)
